refactor(NetworkDiagramBuilder): split into 4 partials, dedupe sizing, add color palette and comments
